Delete buttons for cards and for decks
The deck delete also deletes all associated cards.

+++ website/views.py
@@ -36,23 +36,6 @@ def home():
 
     return render_template("home.html", user=current_user, decks=decks)
 
-    # if request.method == 'POST':
-    #     front = request.form.get('front')
-    #     back = request.form.get('back')
-
-    #     if not front or not back:
-    #         flash('That\'s not much of a flashcard!', category='error')
-    #     else:
-    #         #create the note with the text and the user id
-    #         new_flashcard = Flashcard(front=front, back=back, user_id=current_user.id)
-    #         #add the note do the database
-    #         db.session.add(new_flashcard)
-    #         db.session.commit()
-    #         flash('Flashcard added!', category='success')
-    # #This renders the template and also passes in the current_user variable.
-    # #We can access the associated flashcards with current_user.flashcards
-    # return render_template("home.html", user=current_user)
-
 @views.route('/study/<int:deck_id>', methods=['GET', 'POST'])
 @login_required
 def study(deck_id):
@@ -82,13 +65,37 @@ def view_deck(deck_id):
     return render_template('view_deck.html', deck=deck, deck_id=deck_id, flashcards=flashcards)
 
 @views.route('/delete-flashcard', methods=['POST'])
+@login_required
 def delete_flashcard():
-    flashcard = json.loads(request.data)
-    flashcardId = flashcard['flashcardId']
-    flashcard = Flashcard.query.get(flashcardId)
-    if flashcard: 
-        if flashcard.user_id == current_user.id:
+    #This converts the json data to a python dictionary
+    flashcard_data = json.loads(request.data)
+    #This gets the flashcard id from the dictionary. The key is 'flashcardId' and the value is the acutal flashcard id
+    flashcard_id = flashcard_data['flashcardId']
+    #This gets the flashcard object using the flashcard id
+    flashcard = Flashcard.query.get(flashcard_id)
+
+    #This checks if the current user is the user who created the flashcard
+    if flashcard and flashcard.user_id == current_user.id:
             db.session.delete(flashcard)
             db.session.commit()
-            
-    return jsonify({})
+            #This returns a json object with the deleted key set to true
+            return jsonify({ 'deleted': True })
+    
+    return jsonify({ 'deleted': False })
+
+# This works the same as the delete-flashcard function but also deletes the flashcards in the deck using the deck_id
+@views.route('/delete-deck', methods=['POST'])
+@login_required
+def delete_deck():
+    deck_data = json.loads(request.data)
+    deck_id = deck_data['deckId']
+    deck = Deck.query.get(deck_id)
+
+    if deck and deck.user_id == current_user.id:
+        # Delete all flashcards associated with the deck
+        Flashcard.query.filter_by(deck_id=deck_id).delete()
+        db.session.delete(deck)
+        db.session.commit()
+        return jsonify({'success': True})
+
+    return jsonify({'success': False}), 400
